+++ src/index.js
@@ -1,3 +1,5 @@
+import renderResult from './modules/dom';
+
 /* eslint-disable prefer-destructuring */
 const form = document.querySelector('.top-banner form');
 const input = document.querySelector('.top-banner input');
@@ -51,21 +53,7 @@ form.addEventListener('submit', (e) => {
       } = data;
       const icon = `https://s3-us-west-2.amazonaws.com/s.cdpn.io/162656/${weather[0].icon}.svg`;
 
-      const li = document.createElement('li');
-      li.classList.add('city');
-      const markup = `
-        <h2 class="city-name" data-name="${name},${sys.country}">
-          <span>${name}</span>
-          <sup>${sys.country}</sup>
-        </h2>
-        <div class="city-temp">${Math.round(main.temp)}<sup>°C</sup></div>
-        <figure>
-          <img class="city-icon" src="${icon}" alt="${weather[0].description}">
-          <figcaption>${weather[0].description}</figcaption>
-        </figure>
-      `;
-      li.innerHTML = markup;
-      list.appendChild(li);
+      renderResult(main, name, sys, weather, icon);
     })
     .catch(() => {
       msg.textContent = 'Please search for a valid city 😩';
